+++ app/actions/index.js
@@ -2,22 +2,22 @@
 
 import type { Sensors, Dispatch, GetState } from "../utils/flowtype";
 
-export const SWITCH_BACKEND = 'SWITCH_BACKEND'
+export const SWITCH_BACKEND = 'SWITCH_BACKEND';
 export const switchBackend = (selected:string) => {
     return {
         type: SWITCH_BACKEND,
         selected
     }
-}
+};
 
-export const ADD_ENDPOINTS = 'ADD_ENDPOINTS'
+export const ADD_ENDPOINTS = 'ADD_ENDPOINTS';
 export const addEndpoints = () =>{
     return {
         type: ADD_ENDPOINTS
     }
-}
+};
 
-export const REQUEST_SENSORS = 'REQUEST_SENSORS'
+export const REQUEST_SENSORS = 'REQUEST_SENSORS';
 function requestSensors(api:string) {
     return {
         type: REQUEST_SENSORS,
@@ -25,7 +25,7 @@ function requestSensors(api:string) {
     }
 }
 
-export const RECEIVE_SENSORS = 'RECEIVE_SENSORS'
+export const RECEIVE_SENSORS = 'RECEIVE_SENSORS';
 function receiveSensors(api:string, json:Sensors) {
     return (dispatch:Dispatch) => {
         dispatch({
@@ -39,7 +39,7 @@ function receiveSensors(api:string, json:Sensors) {
     }
 }
 
-export const RECEIVE_SENSOR = 'RECEIVE_SENSOR'
+export const RECEIVE_SENSOR = 'RECEIVE_SENSOR';
 function receiveSensor(json:Sensors) {
     return (dispatch:Dispatch) => {
         dispatch({
@@ -49,7 +49,7 @@ function receiveSensor(json:Sensors) {
     }
 }
 
-export const UPDATE_DETAIL = 'UPDATE_DETAIL'
+export const UPDATE_DETAIL = 'UPDATE_DETAIL';
 export function updateDetail(id:string, name:string, coordinates:number[]){
     return (dispatch:Dispatch) => {
         dispatch({
@@ -61,7 +61,7 @@ export function updateDetail(id:string, name:string, coordinates:number[]){
     }
 }
 
-export const ADD_SEARCH_FILTER = 'ADD_SEARCH_FILTER'
+export const ADD_SEARCH_FILTER = 'ADD_SEARCH_FILTER';
 // this is not used for now, it push a new field in state.searchFilters.filters, not state.searchFilters.selected
 function addSearchFilter(filter:string) {
     return {
@@ -70,7 +70,7 @@ function addSearchFilter(filter:string) {
     }
 }
 
-export const ADD_SEARCH_PARAMETER = 'ADD_SEARCH_PARAMETER'
+export const ADD_SEARCH_PARAMETER = 'ADD_SEARCH_PARAMETER';
 export function addSearchParameter(parameter:Array<string>) {
     return (dispatch:Dispatch, getState:GetState) => {
         const state = getState();
@@ -79,7 +79,7 @@ export function addSearchParameter(parameter:Array<string>) {
             type: ADD_SEARCH_PARAMETER,
             parameter,
             selected_filters
-        })
+        });
         const idx = selected_filters.indexOf('parameters');
         dispatch(updateAvailableSensors(idx));
     }
@@ -97,7 +97,15 @@ export function fetchTrends(parameter:string, totalyear:number, interval:number)
         const season = undefined;
         //TODO: change to year/semi for GLM, may need a config
         const trendsendpoint = api + '/api/geostreams/datapoints/trends?binning=year&semi=all';
-        state.sensors.data.filter(s => s.parameters.indexOf(parameter) >= 0)
+
+        let sensorsToFilter;
+        if ((state.sensorTrends.available_sensors).length > 0 ) {
+            sensorsToFilter = state.sensorTrends.available_sensors;
+        } else {
+            sensorsToFilter = state.sensors.data;
+        }
+
+        sensorsToFilter.filter(s => s.parameters.indexOf(parameter) >= 0)
             .map(sensor => {
                 let start_time = new Date(sensor.min_start_time);
                 let end_time = new Date(sensor.max_end_time);
@@ -189,7 +197,7 @@ export function fetchTrends(parameter:string, totalyear:number, interval:number)
 
 }
 
-export const ADD_SEARCH_DATASOURCE = 'ADD_SEARCH_DATASOURCE'
+export const ADD_SEARCH_DATASOURCE = 'ADD_SEARCH_DATASOURCE';
 export function addSearchDataSource(data_source:Array<string>) {
     return (dispatch:Dispatch, getState:GetState) => {
         const state = getState();
@@ -198,13 +206,13 @@ export function addSearchDataSource(data_source:Array<string>) {
             type: ADD_SEARCH_DATASOURCE,
             data_source,
             selected_filters
-        })
+        });
         const idx = selected_filters.indexOf('data_source');
         dispatch(updateAvailableSensors(idx));
     }
 }
 
-export const ADD_START_DATE = 'ADD_START_DATE'
+export const ADD_START_DATE = 'ADD_START_DATE';
 export function addStartDate(date:Date) {
     return (dispatch:Dispatch, getState:GetState) => {
         const state = getState();
@@ -215,13 +223,13 @@ export function addStartDate(date:Date) {
             date,
             selected_filters,
             availableSensors
-        })
+        });
         const idx = selected_filters.indexOf('time');
         dispatch(updateAvailableSensors(idx));
     }
 }
 
-export const ADD_END_DATE = 'ADD_END_DATE'
+export const ADD_END_DATE = 'ADD_END_DATE';
 export function addEndDate(date:Date) {
     return (dispatch:Dispatch, getState:GetState) => {
         const state = getState();
@@ -232,13 +240,13 @@ export function addEndDate(date:Date) {
             date,
             selected_filters,
             availableSensors
-        })
+        });
         const idx = selected_filters.indexOf('time');
         dispatch(updateAvailableSensors(idx));
     }
 }
 
-export const ADD_SEARCH_LOCATION = 'ADD_SEARCH_LOCATION'
+export const ADD_SEARCH_LOCATION = 'ADD_SEARCH_LOCATION';
 export function addSearchLocation(location:? string) {
     return (dispatch:Dispatch, getState:GetState) => {
         const state = getState();
@@ -247,13 +255,40 @@ export function addSearchLocation(location:? string) {
             type: ADD_SEARCH_LOCATION,
             location,
             selected_filters
-        })
+        });
         const idx = selected_filters.indexOf('locations');
         dispatch(updateAvailableSensors(idx));
     }
 }
 
-export const ADD_FILTER = 'ADD_FILTER'
+export const ADD_CUSTOM_LOCATION_FILTER = 'ADD_CUSTOM_LOCATION_FILTER';
+export function addCustomLocationFilter(selectedPointsLocations:Array<string>) {
+    return (dispatch:Dispatch) => {
+        dispatch({
+            type: ADD_CUSTOM_LOCATION_FILTER,
+            selectedPointsLocations,
+        })
+
+    }
+}
+
+export const ADD_CUSTOM_TREND_LOCATION_FILTER = 'ADD_CUSTOM_TREND_LOCATION_FILTER';
+export function addCustomTrendLocationFilter(selectedPointsLocations:Array<string>) {
+    return (dispatch:Dispatch, getState:GetState) => {
+
+        const state = getState();
+        let origSensors = state.sensors.data;
+
+        dispatch({
+            type: ADD_CUSTOM_TREND_LOCATION_FILTER,
+            selectedPointsLocations,
+            origSensors,
+        })
+
+    }
+}
+
+export const ADD_FILTER = 'ADD_FILTER';
 export function addFilter(selectedFilter:string) {
     return (dispatch:Dispatch) => {
         dispatch({
@@ -264,7 +299,7 @@ export function addFilter(selectedFilter:string) {
     }
 }
 
-export const CHANGE_FILTER = 'CHANGE_FILTER'
+export const CHANGE_FILTER = 'CHANGE_FILTER';
 export function changeFilter(selectedFilter:string, idx:number) {
     return (dispatch:Dispatch) => {
         dispatch({
@@ -277,14 +312,14 @@ export function changeFilter(selectedFilter:string, idx:number) {
     }
 }
 
-export const DELETE_FILTER = 'DELETE_FILTER'
+export const DELETE_FILTER = 'DELETE_FILTER';
 export function deleteFilter(idx:number) {
     return (dispatch:Dispatch) => {
         dispatch(updateAvailableSensors(idx - 1));
     }
 }
 
-export const UPDATE_AVAILABLE_FILTERS = 'UPDATE_AVAILABLE_FILTERS'
+export const UPDATE_AVAILABLE_FILTERS = 'UPDATE_AVAILABLE_FILTERS';
 export function updateAvailableFilters() {
     return (dispatch:Dispatch, getState:GetState) => {
         const state = getState();
@@ -302,8 +337,8 @@ export function updateAvailableFilters() {
     }
 }
 
-export const UPDATE_AVAILABLE_SENSORS = 'UPDATE_AVAILABLE_SENSORS'
-export const DELETE_FILTERS_AFTER = 'DELETE_FILTERS_AFTER'
+export const UPDATE_AVAILABLE_SENSORS = 'UPDATE_AVAILABLE_SENSORS';
+export const DELETE_FILTERS_AFTER = 'DELETE_FILTERS_AFTER';
 export function updateAvailableSensors(idx:number) {
     return (dispatch:Dispatch, getState:GetState) => {
         const state = getState();
@@ -324,7 +359,7 @@ export function updateAvailableSensors(idx:number) {
     }
 }
 
-export const SELECT_SENSOR = 'SELECT_SENSOR'
+export const SELECT_SENSOR = 'SELECT_SENSOR';
 export function selectSensorDetail(id:string, name:string, coordinates:Array<number>) {
     return {
         type: SELECT_SENSOR,
@@ -374,11 +409,11 @@ export function fetchSensor(name:string) {
 
             } else {
                 const endpointsensors = api + '/api/geostreams/sensors';
-                var result = fetch(endpointsensors)
+                let result = fetch(endpointsensors)
                     .then(response => response.json())
                     .then(json => {
                         const sensor = json.find(x => x.name === name);
-                        console.log(sensor.id)
+                        console.log(sensor.id);
                         dispatch(updateDetail(sensor.id, sensor.name, sensor.geometry.coordinates.slice(0, 2)));
                         return fetch(api + '/api/geostreams/datapoints/bin/semi/1?sensor_id=' + sensor.id)
                     });