+++ app/reducers/sensors.js
@@ -1,15 +1,23 @@
 /*
  * @flow
  */
-import { SWITCH_BACKEND, REQUEST_SENSORS, RECEIVE_SENSORS, UPDATE_AVAILABLE_SENSORS} from '../actions'
-import type { Sensor, Sensors, sensorsState, MapWithLabel, MapWithLabels } from '../utils/flowtype'
+import { RECEIVE_SENSORS, UPDATE_AVAILABLE_SENSORS, ADD_CUSTOM_LOCATION_FILTER} from '../actions'
+import type { Sensor,Sensors, sensorsState, MapWithLabel, MapWithLabels } from '../utils/flowtype'
 import {inArray, sortByLabel, pnpoly} from '../utils/arrayUtils'
 import {getLocationName, getSourceName, getParameterName} from '../utils/getConfig'
 
-type SensorAction = {| type:'RECEIVE_SENSORS' | 'UPDATE_AVAILABLE_SENSORS', sensors:Sensors, api:string, receivedAt:Date,
+type SensorAction = {| type:'RECEIVE_SENSORS' | 'UPDATE_AVAILABLE_SENSORS',
+    sensors:Sensors, api:string, receivedAt:Date,
     selected_search:Object, selected_filters:Array<string>|};
 
-const defaultState = {data:[], parameters: [], sources: [], locations:[], available_sensors:[]}
+const defaultState = {
+    data:[],
+    parameters: [],
+    sources: [],
+    locations:[],
+    available_sensors:[],
+    draw_available_sensors:[]
+};
 
 const sensors = (state:sensorsState = defaultState, action:SensorAction) => {
 	switch(action.type) {
@@ -22,26 +30,59 @@ const sensors = (state:sensorsState = defaultState, action:SensorAction) => {
 		        sources: collectSources(action.sensors),
                 locations: collectLocations(action.sensors),
                 available_sensors: action.sensors
-    	})
+    	});
 
-      case UPDATE_AVAILABLE_SENSORS:
-            let newSensors = filterAvailableSensors(state, action.selected_filters, action.selected_search)
+        case UPDATE_AVAILABLE_SENSORS:
+            let newSensors = filterAvailableSensors(state, action.selected_filters, action.selected_search);
             return Object.assign({}, state, {available_sensors: newSensors});
-		default:
-			return state
+
+        case ADD_CUSTOM_LOCATION_FILTER:
+            let customLocationSensors = filterCustomLocation(state, action.selectedPointsLocations);
+            return Object.assign({}, state, {
+                available_sensors: customLocationSensors,
+                draw_available_sensors: customLocationSensors
+            });
+
+        default:
+            return state
 	}
+};
+
+function filterCustomLocation(state:sensorsState, selectedPointsLocations:Array<string>) {
+    let origSensors = state.data;
+    let filteredSensors = [];
+
+    if ( state.available_sensors.length < state.draw_available_sensors.length) {
+        origSensors = state.available_sensors;
+    }
+
+    if (selectedPointsLocations[0] == 'reset_points') {
+
+        filteredSensors = origSensors;
+
+    } else {
+
+        origSensors.map((sensor) => {
+            if (selectedPointsLocations.includes(sensor.name)) {
+                filteredSensors.push(sensor);
+            }
+        });
+
+    }
+
+    return filteredSensors;
 }
 
 export function collectParameters(sensorsData:Sensors):MapWithLabels {
-    var params:MapWithLabels = [];
+    let params:MapWithLabels = [];
     sensorsData.map(s => {
       s.parameters.map(p => {
         // check if parameters exists already
         const found = params.some(function (e:MapWithLabel) {
           return e.id === p;
-        })
+        });
         if (p === null)
-        	console.log(`Found sensor ${s.id} with null parameters`)
+        	console.log(`Found sensor ${s.id} with null parameters`);
         else if (!found && getParameterName(p) != null)
         	params.push({'id': p, 'label': getParameterName(p) || ''});
       });
@@ -51,15 +92,15 @@ export function collectParameters(sensorsData:Sensors):MapWithLabels {
   }
 
 export function collectSources(sensorsData:Sensors):MapWithLabels {
-    var sources:MapWithLabels = [];
+    let sources:MapWithLabels = [];
     sensorsData.map(s => {
-      var source = s.properties.type;
+      let source = s.properties.type;
       // check if source exists already
         const found = sources.some(function (e:MapWithLabel) {
             return e.id === source.id;
-        })
+        });
        if (source === null)
-        console.log(`Found sensor ${s.id} with null data sources`)
+        console.log(`Found sensor ${s.id} with null data sources`);
       else if (!found) 
       	sources.push({'id':source.id, 'label': getSourceName(source) || ''});
     });
@@ -119,6 +160,15 @@ export function collectLocations(sensorsData:Sensors):MapWithLabels {
 function filterAvailableSensors(state:sensorsState, selectedFilters:Array<string>, selectedSearch:Object) {
     let av_sensors: Sensors = state.data;
     let new_sensors: Sensors = [];
+    let draw_sensors = state.draw_available_sensors;
+
+    let draw_sensors_names = [];
+    for (let i = 0; i < draw_sensors.length; i++) {
+        if (!draw_sensors_names.includes((draw_sensors[i].name).toString())) {
+            draw_sensors_names.push(draw_sensors[i]);
+        }
+    }
+
     selectedFilters.map ((filter) => {
         switch(filter) {
             case 'data_source':
@@ -150,7 +200,8 @@ function filterAvailableSensors(state:sensorsState, selectedFilters:Array<string
             case 'time':
                 new_sensors=[];
                 av_sensors.map((sensor) => {
-                    if(selectedSearch.dates.selected.start <= new Date(sensor.max_end_time) && selectedSearch.dates.selected.end >= new Date(sensor.min_start_time)) {
+                    if(selectedSearch.dates.selected.start <= new Date(sensor.max_end_time) &&
+                        selectedSearch.dates.selected.end >= new Date(sensor.min_start_time)) {
                         new_sensors.push(sensor);
                     }
                 });
@@ -160,11 +211,22 @@ function filterAvailableSensors(state:sensorsState, selectedFilters:Array<string
             case 'locations':
                 if(selectedSearch.locations.selected != null ) {
                     new_sensors = [];
-                    av_sensors.map((sensor => {
-                        if(matchLocation(selectedSearch.locations.selected, sensor)) {
-                            new_sensors.push(sensor);
+
+                    if (selectedSearch.locations.selected == 'draw') {
+                        if (draw_sensors_names.length <= 0) {
+                            new_sensors = av_sensors;
+                        } else {
+                            new_sensors = av_sensors.filter(function(e){return this.indexOf(e)>=0;},draw_sensors_names);
                         }
-                    }));
+                    }
+                    else {
+                        av_sensors.map((sensor => {
+                            if(matchLocation(selectedSearch.locations.selected, sensor)) {
+                                new_sensors.push(sensor);
+                            }
+                        }));
+                    }
+
                     av_sensors = new_sensors;
                 }